refactor(webapp/explore): migrate Explore page to tailwindcss
